imaging code initial
The code I use to analyze fly brain calcium activity (collected between
2 and 6 Hz) and the fly’s behaviour (collected at 50 Hz).

Main use of `imaging` is to create a GCaMP_ABF object. The arguments to
this are a path to a folder (str) containing a volumetric movie of
brain activity (as a .tiff) and an “axon binary file” annotating the
fly’s movements (.abf)

Roughly, the hierarchy is `imaging` calls  `behaviour`, `circstats`, and `functions` . `behaviour` also calls `functions`.